Added Firestore listener.
Made the program more modular added the firestore listener to listen for active sports for betting platforms as well as tuned some stuff in the realtime listener. Lastly, Added a dispatcher that controls threads for sports monitoring. When sport is active it spools up a thread for the sport otherwise if it goes inactive it will stop thread and free up space for other sports