Fixed prizepicks{SportsName}
So, I was getting and issue where the player information wasn't persisting though process projections when image_url wasn't present we marked it for remaining so we have everyone's picture. This was overwriting existing data and in the end we would only have the image_url and projections merged together. We fixed that issue and currently montioring if this issue happens again